+++ neoforge/src/main/java/software/bluelib/api/event/mod/ModLoadedEvent.java
@@ -1,7 +1,5 @@
 package software.bluelib.api.event.mod;
 
-import java.net.URL;
-import java.util.Map;
 import java.util.Optional;
 import net.neoforged.bus.api.Event;
 import net.neoforged.fml.event.IModBusEvent;
@@ -40,25 +38,6 @@ public class ModLoadedEvent extends Event implements IModBusEvent {
 
     public String getDescription() {
         return modData.description();
-    }
-
-    @NotNull
-    public String getNamespace() {
-        return modData.namespace();
-    }
-
-    public Map<String, Object> getProperties() {
-        return modData.properties();
-    }
-
-    @Nullable
-    public Optional<URL> getUpdateURL() {
-        return modData.updateURL();
-    }
-
-    @Nullable
-    public Optional<URL> getModURL() {
-        return modData.modURL();
     }
 
     @Nullable
