+++ neoforge/src/main/java/software/bluelib/example/event/ReloadHandler.java
@@ -12,7 +12,7 @@ import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.AddReloadListenerEvent;
 import net.neoforged.neoforge.event.server.ServerStartingEvent;
 import software.bluelib.BlueLibConstants;
-import software.bluelib.event.ReloadEventHandler;
+import software.bluelib.api.reload.ReloadEventHandler;
 import software.bluelib.utils.logging.BaseLogLevel;
 import software.bluelib.utils.logging.BaseLogger;
